Update UI and improve timer logic

+++ app/src/main/java/com/example/android/allocate/activities/MainActivity.java
@@ -1,12 +1,7 @@
 package com.example.android.allocate.activities;
 
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.os.SystemClock;
 import android.support.design.widget.FloatingActionButton;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -40,14 +35,14 @@ public class MainActivity extends AppCompatActivity {
     private Toolbar toolbar;
 
     private TaskHandler mTaskHandler;
-    private TickReciever  mBroadcastReceiver;
+    private TickReciever mBroadcastReceiver;
 
     private TimerDoneReceiver timerDoneReceiver = new TimerDoneReceiver();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.fragment_main);
+        setContentView(R.layout.activity_main);
 
         mTaskHandler = new TaskHandler(this);
 
@@ -74,8 +69,12 @@ public class MainActivity extends AppCompatActivity {
         ItemTouchHelper.SimpleCallback simpleItemTouchCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
             @Override
             public void onSwiped(RecyclerView.ViewHolder viewHolder, int swipeDir) {
-                mTaskHandler.getTaskAdapter().deleteItem(viewHolder.getAdapterPosition());
-                Toast.makeText(MainActivity.this, "Task Deleted", Toast.LENGTH_SHORT).show();
+                if (mTaskHandler.getDataset().get(viewHolder.getAdapterPosition()).isRunning())
+                    return;
+                else {
+                    mTaskHandler.getTaskAdapter().deleteItem(viewHolder.getAdapterPosition());
+                    Toast.makeText(MainActivity.this, "Task Deleted", Toast.LENGTH_SHORT).show();
+                }
             }
 
             @Override
@@ -84,7 +83,7 @@ public class MainActivity extends AppCompatActivity {
                 return true;
             }
 
-            
+
         };
 
         ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleItemTouchCallback);
@@ -101,7 +100,11 @@ public class MainActivity extends AppCompatActivity {
         mBroadcastReceiver = new TickReciever();
         mBroadcastReceiver.initialize(mTaskHandler);
 
-        timerDoneReceiver.cancelAlarms(this);
+        for (Task t : mTaskHandler.getDataset()) {
+            if (t.isRunning()) {
+                timerDoneReceiver.cancelAlarms(this);
+            }
+        }
 
         Log.i(TimerBroadcastService.COUNTDOWN_BROADCAST, "Started service");
         registerReceiver(mBroadcastReceiver, new IntentFilter(TimerBroadcastService.COUNTDOWN_BROADCAST));
@@ -115,9 +118,9 @@ public class MainActivity extends AppCompatActivity {
         stopService(new Intent(this, TimerBroadcastService.class));
         Log.i(TimerBroadcastService.COUNTDOWN_BROADCAST, "Stopped service");
 
-        for(Task t : mTaskHandler.getDataset()){
-            if(t.isRunning()){
-                timerDoneReceiver.setAlarms(this,t.getTimeLeft(),t.getId());
+        for (Task t : mTaskHandler.getDataset()) {
+            if (t.isRunning()) {
+                timerDoneReceiver.setAlarms(this, t.getTimeLeft(), t.getId());
             }
         }
 
@@ -144,10 +147,10 @@ public class MainActivity extends AppCompatActivity {
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
 
-        switch(item.getItemId()){
-            case R.id.action_settings :
+        switch (item.getItemId()) {
+            case R.id.action_settings:
                 break;
-            case R.id.action_clear :
+            case R.id.action_clear:
                 mTaskHandler.clearTask();
                 Toast.makeText(MainActivity.this, "Task Cleared", Toast.LENGTH_SHORT).show();
                 break;
@@ -157,7 +160,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void startAddTaskActivity() {
-        Intent intent  = new Intent(this, AddTaskActivity.class);
+        Intent intent = new Intent(this, AddTaskActivity.class);
         startActivity(intent);
     }
 }
